requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
langchain-groq==0.1.1
langchain-huggingface==0.0.12
langchain-community==0.0.27
langchain-core==0.1.27
langchain==0.1.9
pypdf==4.0.2
faiss-cpu==1.7.4
sentence-transformers==2.5.1 